+++ src/omnisharp/requirementCheck.ts
@@ -54,7 +54,7 @@ export async function validateRequirements(): Promise<boolean> {
 async function checkRequirements(): Promise<RequirementResult> {
     if (omnisharpOptions.useModernNet) {
         const dotnetInfo = await getDotnetInfo(omnisharpOptions.dotNetCliPaths);
-        const needsDotNetSdk = dotnetInfo.Version === undefined || semver.lt(dotnetInfo.Version, '6.0.0');
+        const needsDotNetSdk = dotnetInfo.Version === undefined || semver.lt(dotnetInfo.Version, '10.0.0');
         return {
             needsDotNetSdk,
             needsMono: false,