+++ test/omnisharp/omnisharpUnitTests/optionChangeObserver.test.ts
@@ -37,7 +37,6 @@ describe('OmniSharpConfigChangeObserver', () => {
         { config: 'omnisharp', section: 'path', value: 'somePath' },
         { config: 'omnisharp', section: 'waitForDebugger', value: true },
         { config: 'omnisharp', section: 'enableMsBuildLoadProjectsOnDemand', value: true },
-        { config: 'omnisharp', section: 'useModernNet', value: false },
         { config: 'omnisharp', section: 'loggingLevel', value: 'verbose' },
     ].forEach((elem) => {
         describe(`When the ${elem.config} ${elem.section} changes`, () => {